ページのマージ作業
