+++ src/app/pages/main.module.scss
@@ -1,0 +1,9 @@
+.mainContainer {
+  width: 100%;
+
+  .section {
+    min-height: 100vh;
+    scroll-margin-top: 60px;
+    width: 100%;
+  }
+}
