+++ src/app/pages/sections/mission/component/mission-no/mission-no.module.scss
