+++ src/app/pages/sections/service/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import '../globals.css';
-import { ServiceAndRecruit } from '@/app/shared/serviceAndRecruit/serviceAndRecruit';
+import '@app/globals.css';
+import { ServiceAndRecruit } from '@app/shared/serviceAndRecruit/serviceAndRecruit';
 
 const Service: React.FC = () => {
   return <ServiceAndRecruit></ServiceAndRecruit>;
