+++ src/app/pages/sections/top/components/header/logo-icon/logo-icon.tsx
