--- src/app/pages/top/components/header/logo-icon/logo-icon.module.scss
@@ -1,4 +1,0 @@
-.logoIcon {
-  font-size: 20px;
-  font-weight: bold;
-}
