+++ src/app/security/page.tsx
@@ -1,9 +1,9 @@
 // 'use-strict';
 'use client';
 
-import MonochromeFooter from '@/app/components/footer/monochrome-footer/monochrome-footer';
+import MonochromeFooter from '@app/components/footer/monochrome-footer/monochrome-footer';
+import { useEffect, useState } from 'react';
 import styles from './security.module.scss';
-import React, { useEffect, useState } from 'react';
 
 const Security = () => {
   const [data, setData] = useState<string | null>(null);
