+++ src/app/shared/serviceAndRecruit/serviceAndRecruit.tsx
@@ -1,8 +1,7 @@
 'use client';
 
-import ColorFooter from '@/app/components/footer/color-footer/color-footer';
-import { Recruit } from '@/app/shared/serviceAndRecruit/components/recruit-page/recruit';
-import { Service } from '@/app/shared/serviceAndRecruit/components/service-page/service';
+import { Recruit } from '@app/shared/serviceAndRecruit/components/recruit-page/recruit';
+import { Service } from '@app/shared/serviceAndRecruit/components/service-page/service';
 import '@i18n';
 import styles from './serviceAndRecruit.module.scss';
 
@@ -11,9 +10,9 @@ export const ServiceAndRecruit: React.FC = () => {
     <div className={styles.servicePage}>
       <Service></Service>
       <Recruit></Recruit>
-      <footer>
+      {/* <footer>
         <ColorFooter backGround=""></ColorFooter>
-      </footer>
+      </footer> */}
     </div>
   );
 };
